requirements.txt
aiohttp==3.11.10
asyncpraw==7.8.1
datasets==3.6.0
discord.py==2.5.2
faiss_cpu==1.11.0
feedparser==6.0.11
huggingface_hub==0.26.5
numpy==2.3.1
peft==0.15.2
PixivPy_Async==1.2.14
PyPDF2==3.0.1
python-dotenv==1.1.1
pytz==2024.2
Requests==2.32.4
sentence_transformers==4.1.0
torch==2.5.1+cu124
transformers==4.47.0
yt_dlp==2025.6.30
